+++ np_nurbs/rational_bezier.py
@@ -10,23 +10,23 @@ __all__ = [
 
 
 def rational_bezier_curve_eval_grid(
-        p: np.ndarray[tuple[int, int], np.dtype[float]], 
-        w: np.ndarray[tuple[int,], np.dtype[float]],
+        p: np.ndarray[tuple[int, int], np.dtype[np.float64]], 
+        w: np.ndarray[tuple[int,], np.dtype[np.float64]],
         nt: int,
-        ) -> np.ndarray[tuple[int, int], np.dtype[float]]:
+        ) -> np.ndarray[tuple[int, int], np.dtype[np.float64]]:
     """
     Evaluates a rational Bézier curve on an evenly spaced parameter vector
     (``linspace(0, 1, nt)``) using a fully vectorized formulation.
 
     Parameters
     ----------
-    p: np.ndarray[tuple[int, int], np.dtype[float]]
+    p: np.ndarray[tuple[int, int], np.dtype[np.float64]]
         Rational Bézier curve control point array. 
         This array has shape
         :math:`(n+1) \\times d`, where :math:`n` is
         the curve degree and :math:`d` is the number
         of dimensions (usually 2 or 3)
-    w: np.ndarray[tuple[int,], np.dtype[float]]
+    w: np.ndarray[tuple[int,], np.dtype[np.float64]]
         Vector of weights, corresponding one-to-one with
         the control points
     nt: int
@@ -35,14 +35,14 @@ def rational_bezier_curve_eval_grid(
 
     Returns
     -------
-    np.ndarray[tuple[int, int], np.dtype[float]]
+    np.ndarray[tuple[int, int], np.dtype[np.float64]]
         The evaluated rational Bézier curve with shape
         :math:`n_t \\times d`, where :math:`n_t`
         is the number of parameters
     """
     assert len(p) == len(w)
     degree = len(p) - 1
-    t = np.linspace(0.0, 1.0, nt, dtype=float)
+    t = np.linspace(0.0, 1.0, nt, dtype=np.float64)
     powers = (degree - np.arange(degree + 1))[:, np.newaxis]
     t_mat = t ** powers
     
@@ -59,11 +59,11 @@ def rational_bezier_curve_eval_grid(
 
 
 def rational_bezier_surf_eval_grid(
-        p: np.ndarray[tuple[int, int, int], np.dtype[float]],
-        w: np.ndarray[tuple[int, int], np.dtype[float]],
+        p: np.ndarray[tuple[int, int, int], np.dtype[np.float64]],
+        w: np.ndarray[tuple[int, int], np.dtype[np.float64]],
         nu: int, 
         nv: int,
-        ) -> np.ndarray[tuple[int, int, int], np.dtype[float]]:
+        ) -> np.ndarray[tuple[int, int, int], np.dtype[np.float64]]:
     """
     Evaluates a rational Bézier surface on a uniform parameter grid
     (``linspace(0, 1, nu), linspace(0, 1, nv)``) 
@@ -71,14 +71,14 @@ def rational_bezier_surf_eval_grid(
 
     Parameters
     ----------
-    p: np.ndarray[tuple[int, int, int], np.dtype[float]]
+    p: np.ndarray[tuple[int, int, int], np.dtype[np.float64]]
         Rational Bézier surface control point array. 
         This array has shape
         :math:`(n+1) \\times (m+1) \\times d`, where :math:`n` is
         the surface degree in the :math:`u`-direction,
         :math:`m` is the surface degree in the :math:`v`-direction,
         and :math:`d` is the number of dimensions (usually 3)
-    w: np.ndarray[tuple[int, int], np.dtype[float]]
+    w: np.ndarray[tuple[int, int], np.dtype[np.float64]]
         Vector of weights, corresponding one-to-one with
         the control points
     nu: int
@@ -90,7 +90,7 @@ def rational_bezier_surf_eval_grid(
 
     Returns
     -------
-    np.ndarray[tuple[int, int, int], np.dtype[float]]
+    np.ndarray[tuple[int, int, int], np.dtype[np.float64]]
         The evaluated rational Bézier surface with shape
         :math:`n_u \\times n_v \\times d`, where :math:`n_u`
         is the number of parameters in the
@@ -99,8 +99,8 @@ def rational_bezier_surf_eval_grid(
     """
     n = p.shape[0] - 1
     m = p.shape[1] - 1
-    u = np.linspace(0.0, 1.0, nu, dtype=float)
-    v = np.linspace(0.0, 1.0, nv, dtype=float)
+    u = np.linspace(0.0, 1.0, nu, dtype=np.float64)
+    v = np.linspace(0.0, 1.0, nv, dtype=np.float64)
     u_powers = (n - np.arange(n + 1))[:, np.newaxis]
     v_powers = (m - np.arange(m + 1))[:, np.newaxis]
     u_mat = u ** u_powers
