+++ tests/test_rational_bezier.py
@@ -10,28 +10,28 @@ import rust_nurbs
 
 
 @pytest.fixture
-def p_curve() -> np.ndarray[tuple[int, int], np.dtype[float]]:
+def p_curve() -> np.ndarray[tuple[int, int], np.dtype[np.float64]]:
     return np.random.uniform(low=-5.0, high=5.0, size=(10, 3))
 
 
 @pytest.fixture
-def p_surf() -> np.ndarray[tuple[int, int, int], np.dtype[float]]:
+def p_surf() -> np.ndarray[tuple[int, int, int], np.dtype[np.float64]]:
     return np.random.uniform(low=-5.0, high=5.0, size=(10, 10, 3))
 
 
 @pytest.fixture
-def w_curve() -> np.ndarray[tuple[int,], np.dtype[float]]:
+def w_curve() -> np.ndarray[tuple[int,], np.dtype[np.float64]]:
     return np.random.uniform(low=0.01, high=10.0, size=(10,))
 
 
 @pytest.fixture
-def w_surf() -> np.ndarray[tuple[int, int], np.dtype[float]]:
+def w_surf() -> np.ndarray[tuple[int, int], np.dtype[np.float64]]:
     return np.random.uniform(low=0.01, high=10.0, size=(10, 10))
 
 
 def test_rational_bezier_curve_eval_grid(
-        p_curve: np.ndarray[tuple[int, int], np.dtype[float]],
-        w_curve: np.ndarray[tuple[int,], np.dtype[float]]
+        p_curve: np.ndarray[tuple[int, int], np.dtype[np.float64]],
+        w_curve: np.ndarray[tuple[int,], np.dtype[np.float64]]
         ):
     np_curve = np_nurbs.rational_bezier_curve_eval_grid(
             p_curve, w_curve, 150
@@ -43,8 +43,8 @@ def test_rational_bezier_curve_eval_grid(
 
 
 def test_rational_bezier_surf_eval_grid(
-        p_surf: np.ndarray[tuple[int, int, int], np.dtype[float]],
-        w_surf: np.ndarray[tuple[int, int], np.dtype[float]]
+        p_surf: np.ndarray[tuple[int, int, int], np.dtype[np.float64]],
+        w_surf: np.ndarray[tuple[int, int], np.dtype[np.float64]]
         ):
     np_surf = np_nurbs.rational_bezier_surf_eval_grid(
             p_surf, w_surf, 50, 50
